+++ views/config/index.php
@@ -24,7 +24,7 @@ $module = Yii::$app->getModule('move-content');
 
         <div class="panel-body">
 
-            <div class="alert alert-info">
+            <div class="alert alert-info cuzy-free-module-info">
                 This module was created and is maintained by
                 <a href="https://www.cuzy.app/"
                    target="_blank">CUZY.APP (view other modules)</a>.