+++ docs/CHANGELOG.md
@@ -6,6 +6,7 @@ Unreleased
 - Enh: When moving content from a user to another, update the attached files ownership too 
 - Enh: When moving content from a user to another, update other tables that are not content
 - Enh: Add GitHub HumHub PHP workflows (tests & CS fixer)
+- Fix: If a module is installed but not enabled, non-content active records can fail to be moved
 
 2.1.3 (Sept 21, 2024)
 --------------------